ch(schema stitching): spliting schimas and merge them into on schema
