+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "dev": "nodemon --exec babel-node index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint index.js",
-    "lint:fix": "eslint --fix index.js"
+    "lint": "eslint src/",
+    "lint:fix": "eslint --fix src/"
   },
   "repository": {
     "type": "git",
