+++ src/models/index.js
@@ -1,0 +1,30 @@
+const users = {
+  1: {
+    id: '1',
+    username: 'Altare Aksanti',
+    messageIds: [1],
+  },
+  2: {
+    id: '2',
+    username: 'Mapendo Celestin',
+    messageId: [2],
+  },
+};
+
+const messages = {
+  1: {
+    id: '1',
+    text: 'Hello World',
+    userId: '1',
+  },
+  2: {
+    id: '2',
+    text: 'By Wolrd',
+    userId: '2',
+  },
+};
+
+export default {
+  users,
+  messages,
+};
