+++ src/resolvers/index.js
@@ -1,0 +1,4 @@
+import userResolvers from './user';
+import messageResolvers from './message';
+
+export default [userResolvers, messageResolvers];
