+++ src/schema/index.js
@@ -1,0 +1,20 @@
+import { gql } from 'apollo-server-express';
+
+import userSchema from './user';
+import messageSchema from './message';
+
+const linkSchema = gql`
+  type Query {
+    _: Boolean
+  }
+
+  type Mutation {
+    _: Boolean
+  }
+
+  type Subscription {
+    _: Boolean
+  }
+`;
+
+export default [linkSchema, userSchema, messageSchema];
