+++ src/schema/message.js
@@ -1,0 +1,19 @@
+import { gql } from 'apollo-server-express';
+
+export default gql`
+    extend type Query {
+        messages: [Message!]!
+        message(id: ID!): message!
+    }
+
+    extend type Mutation {
+        createMessage(text: String!): Message!
+        deleteMessage(id: ID!): Boolean!
+    }
+
+    type Message {
+        id: ID!
+        text: String!
+        user: User!
+    }
+`;
